waf: libtdb: set version in dynlib on OSX
On OSX, without this patch `-compatibility_version` and `-current-version` are
passed to clang during compilationg and not during linking. The resulting
libtdb.dynlib ends up without versions set (`otool -l` reports them as 0.0.0).
I don't think adding these to compile flags is useful at all, clang ignores
them. Not having version set is not fatal, but the current code already
attempts to set it -- this patch is a potential fix to existing behavior.

The rest of the patch is to remove these flags whenever `-dynamiclib` is
removed, because these flags only make sense with it. Unfortunately, this
removal happens in two places, due to pre-existing code duplication -- this
patch just changes both places.

Minor behaviour change, in case it has some unintended sideffects:
`-single_module` flag was only removed if `-dynamiclib` was present; after this
patch, they are both removed separately if present.

I only tested this patch against the *standalone* release of libtdb v1.13.6
from here: https://www.samba.org/ftp/tdb/. I built in the context of the brew
formula for tdb that I am about to submit to Homebrew.

Also, note that that I had to `./configure` with `--disable-rpath
--disable-rpath-install --disable-rpath-private-install`, otherwise
applications do link against the resulting libtdb.dynlib but fail at runtime,
because they are looking for libtdb.inst.dynlib instead of libtdb.dynlib and
not finding it.

Signed-off-by: Alexei Colin <ac@alexeicolin.com>